Refactor projects data
